requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
django-debug-toolbar==3.6.0
idna==3.3
pyTelegramBotAPI==4.7.0
requests==2.28.1
schedule==1.1.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
